pyproject.toml
[tool.poetry]
name = "BLADE"
version = "0.1.0"
description = ""
authors = ["Ken Gu <ken.qgu@gmail.com>"]
readme = "README.md"
packages = [
    { include = "blade_bench" },
]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.7.4"
diskcache = "^5.6.3"
loguru = "^0.7.2"
backoff = "^2.2.1"
openai = "^1.35.3"
absl-py = "^2.1.0"
tiktoken = "^0.7.0"
python-dotenv = "^1.0.1"
anthropic = "^0.29.0"
google-generativeai = "^0.7.0"
langchain = "^0.2.5"
networkx = "^3.3"
hydra-core = "^1.3.2"
pandas = "1.5.3"
astor = "^0.8.1"
rich = "^13.7.1"
nbclient = "^0.10.0"
ipython = "^8.25.0"
pillow = "^10.3.0"
nbformat = "^5.10.4"
ipykernel = "^6.29.4"
scipy = "^1.13.1"
pytest = "^8.2.2"
scikit-learn = "^1.5.0"
statsmodels = "^0.14.2"
matplotlib = "^3.9.0"
pytest-asyncio = "^0.23.7"
transformers = "^4.41.2"
torch = "^2.3.1"
accelerate = "^0.31.0"
fastapi = "^0.111.0"
requests-mock = "^1.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
